requirements.txt
dash>=2.14.0
dash-bootstrap-components>=1.5.0
sqlalchemy>=2.0.0
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
plotly>=5.18.0
